creation of endpoint for requesting vcf quality statistics. Also creating graphics presentation layout

+++ backend/src/main/java/com/ufrpe/vcfanalyzer/repository/FileDBRepository.java
@@ -3,9 +3,9 @@ package com.ufrpe.vcfanalyzer.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.ufrpe.vcfanalyzer.model.FileVcfData;
+import com.ufrpe.vcfanalyzer.domain.FileVcfData;
 
 @Repository
-public interface FileDBRepository extends JpaRepository<FileVcfData, Long> {
+public interface FileDBRepository extends JpaRepository<FileVcfData, Integer> {
 
 }
